Log to stdout. cd to code dir and create DB file there
Till now, logging and DB files were in / if the program runs as a daemon.
Previously DB file was created when the program ran from the command line. To migrate
all data from / to the old file:
* sqlite3 /readings.sqlite
* .mode insert readings
* .out /home/pi/gas_monitor/readings.sql
* .schema ros
* select * from ros;
* .schema readings
* select * from readings;
* .quit
* cd /home/pi/gas_monitor
* sqlite3 -init readings.sql readings.sqlite
* rm readings.sql /readings.sqlite

Change print() to logging.info() to see in logs

Service logs "started"/"stopped" by itself