Use a separate process to retry failed uploads
This prevents delay in uploading new readings after recording but not uploading
a large number of readings, such as a long loss of connectivity.

We create a separate executable python module (that does not create a DB if it
does not exist). The uploader runs as a separate service.

The monitor now uploads each reading independently of recording it and
independently from uploading of the backlog of previously recorded
readings.

Also move erros response handlng into upload()